refactor(CourseManager): import getNextActivityOrModule from ActivityNavigationConditionsManager
