+++ services/ActivityNavigationConditionsManager.php
@@ -153,7 +153,7 @@ class ActivityNavigationConditionsManager
         
         if (in_array($result[self::STATUS_LABEL], [self::STATUS_CODE_OK,self::STATUS_CODE_OK_REACTIONS_NEEDED])) {
             if (is_null($nextCourseStructure)) {
-                $nextCourseStructure = $this->getNextActivityOrModule($data['course'], $data['module'], $data['activity']);
+                $nextCourseStructure = $this->courseManager->getNextActivityOrModule($data['course'], $data['module'], $data['activity']);
             }
             /* check status if all is OK*/
             if (!$nextCourseStructure) {
@@ -302,28 +302,6 @@ class ActivityNavigationConditionsManager
 
         $data[self::STATUS_LABEL] = self::STATUS_CODE_OK ;
         return $data;
-    }
-
-    /** getNextActivityOrModule
-     * @param Course $course
-     * @param Module $module
-     * @param Activity $activity
-     * @return CourseStructure next activity or module
-     */
-    public function getNextActivityOrModule(Course $course, Module $module, Activity $activity): ?CourseStructure
-    {
-        if ($activity->getTag() == $module->getLastActivityTag()) {
-            if ($module->getTag() != $course->getLastModuleTag()) {
-                return $course->getNextModule($module->getTag());
-                // if the current page is the last activity of the module and the module is not the last one,
-                // the next link is to the next module entry
-                // (no next button is showed for the last activity of the last module)
-            }
-        } else {
-            // otherwise, the current activity is not the last of the module and the next link is set to the next activity
-            return $module->getNextActivity($activity->getTag());
-        }
-        return null;
     }
 
     /** checkReactionNeeded
